Module google_adexchangebuyer1d4::api

Structs

Account
Configuration data for an Ad Exchange buyer account.

AccountBidderLocation
Your bidder locations that have distinct URLs.

AccountGetCall
Gets one account by ID.

AccountListCall
Retrieves the authenticated user's list of accounts.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the AdExchangeBuyer hub.

AccountPatchCall
Updates an existing account. This method supports patch semantics.

AccountUpdateCall
Updates an existing account.

AccountsList
An account feed lists Ad Exchange buyer accounts that the user has access to. Each entry in the feed corresponds to a single buyer account.

AdExchangeBuyer
Central instance to access all AdExchangeBuyer related resource activities

AddOrderDealsRequest
There is no detailed description.

AddOrderDealsResponse
There is no detailed description.

AddOrderNotesRequest
There is no detailed description.

AddOrderNotesResponse
There is no detailed description.

BillingInfo
The configuration data for an Ad Exchange billing info.

BillingInfoGetCall
Returns the billing information for one account specified by account ID.

BillingInfoList
A billing info feed lists Billing Info the Ad Exchange buyer account has access to. Each entry in the feed corresponds to a single billing info.

BillingInfoListCall
Retrieves a list of billing information for all accounts of the authenticated user.

BillingInfoMethods
A builder providing access to all methods supported on billingInfo resources. It is not used directly, but through the AdExchangeBuyer hub.

Budget
The configuration data for Ad Exchange RTB - Budget API.

BudgetGetCall
Returns the budget information for the adgroup specified by the accountId and billingId.

BudgetMethods
A builder providing access to all methods supported on budget resources. It is not used directly, but through the AdExchangeBuyer hub.

BudgetPatchCall
Updates the budget amount for the budget of the adgroup specified by the accountId and billingId, with the budget amount in the request. This method supports patch semantics.

BudgetUpdateCall
Updates the budget amount for the budget of the adgroup specified by the accountId and billingId, with the budget amount in the request.

Buyer
There is no detailed description.

ContactInformation
There is no detailed description.

CreateOrdersRequest
There is no detailed description.

CreateOrdersResponse
There is no detailed description.

Creative
A creative and its classification data.

CreativeAdTechnologyProviders
There is no detailed description.

CreativeAddDealCall
Add a deal id association for the creative.

CreativeCorrections
Shows any corrections that were applied to this creative. Read-only. This field should not be set in requests.

CreativeCorrectionsContexts
All known serving contexts containing serving status information.

CreativeDealIds
The external deal ids associated with a creative.

CreativeDealIdsDealStatuses
A list of external deal ids and ARC approval status.

CreativeFilteringReasons
The filtering reasons for the creative. Read-only. This field should not be set in requests.

CreativeFilteringReasonsReasons
The filtering reasons.

CreativeGetCall
Gets the status for a single creative. A creative will be available 30-40 minutes after submission.

CreativeInsertCall
Submit a new creative.

CreativeListCall
Retrieves a list of the authenticated user's active creatives. A creative will be available 30-40 minutes after submission.

CreativeListDealCall
Lists the external deal ids associated with the creative.

CreativeMethods
A builder providing access to all methods supported on creative resources. It is not used directly, but through the AdExchangeBuyer hub.

CreativeNativeAd
If nativeAd is set, HTMLSnippet, videoVastXML, and the videoURL outside of nativeAd should not be set. (The videoURL inside nativeAd can be set.)

CreativeNativeAdAppIcon
The app icon, for app download ads.

CreativeNativeAdImage
A large image.

CreativeNativeAdLogo
A smaller image, for the advertiser logo.

CreativeRemoveDealCall
Remove a deal id associated with the creative.

CreativeServingRestrictions
The granular status of this ad in specific contexts. A context here relates to where something ultimately serves (for example, a physical location, a platform, an HTTPS vs HTTP request, or the type of auction). Read-only. This field should not be set in requests. See the examples in the Creatives guide for more details.

CreativeServingRestrictionsContexts
All known contexts/restrictions.

CreativeServingRestrictionsDisapprovalReasons
The reasons for disapproval within this restriction, if any. Note that not all disapproval reasons may be categorized, so it is possible for the creative to have a status of DISAPPROVED or CONDITIONALLY_APPROVED with an empty list for disapproval_reasons. In this case, please reach out to your TAM to help debug the issue.

CreativesList
The creatives feed lists the active creatives for the Ad Exchange buyer accounts that the user has access to. Each entry in the feed corresponds to a single creative.

DealServingMetadata
There is no detailed description.

DealServingMetadataDealPauseStatus
Tracks which parties (if any) have paused a deal. The deal is considered paused if has_buyer_paused || has_seller_paused. Each of the has_buyer_paused or the has_seller_paused bits can be set independently.

DealTerms
There is no detailed description.

DealTermsGuaranteedFixedPriceTerms
There is no detailed description.

DealTermsGuaranteedFixedPriceTermsBillingInfo
There is no detailed description.

DealTermsNonGuaranteedAuctionTerms
There is no detailed description.

DealTermsNonGuaranteedFixedPriceTerms
There is no detailed description.

DealTermsRubiconNonGuaranteedTerms
There is no detailed description.

DeleteOrderDealsRequest
There is no detailed description.

DeleteOrderDealsResponse
There is no detailed description.

DeliveryControl
There is no detailed description.

DeliveryControlFrequencyCap
There is no detailed description.

Dimension
This message carries publisher provided breakdown. E.g. {dimension_type: 'COUNTRY', [{dimension_value: {id: 1, name: 'US'}}, {dimension_value: {id: 2, name: 'UK'}}]}

DimensionDimensionValue
Value of the dimension.

EditAllOrderDealsRequest
There is no detailed description.

EditAllOrderDealsResponse
There is no detailed description.

GetOffersResponse
There is no detailed description.

GetOrderDealsResponse
There is no detailed description.

GetOrderNotesResponse
There is no detailed description.

GetOrdersResponse
There is no detailed description.

GetPublisherProfilesByAccountIdResponse
There is no detailed description.

MarketplaceDeal
A proposal can contain multiple deals. A deal contains the terms and targeting information that is used for serving.

MarketplaceDealParty
There is no detailed description.

MarketplaceLabel
There is no detailed description.

MarketplaceNote
A proposal is associated with a bunch of notes which may optionally be associated with a deal and/or revision number.

MarketplacedealDeleteCall
Delete the specified deals from the proposal

MarketplacedealInsertCall
Add new deals for the specified proposal

MarketplacedealListCall
List all the deals for a given proposal

MarketplacedealMethods
A builder providing access to all methods supported on marketplacedeal resources. It is not used directly, but through the AdExchangeBuyer hub.

MarketplacedealUpdateCall
Replaces all the deals in the proposal with the passed in deals

MarketplacenoteInsertCall
Add notes to the proposal

MarketplacenoteListCall
Get all the notes associated with a proposal

MarketplacenoteMethods
A builder providing access to all methods supported on marketplacenote resources. It is not used directly, but through the AdExchangeBuyer hub.

MarketplaceprivateauctionMethods
A builder providing access to all methods supported on marketplaceprivateauction resources. It is not used directly, but through the AdExchangeBuyer hub.

MarketplaceprivateauctionUpdateproposalCall
Update a given private auction proposal

MobileApplication
There is no detailed description.

PerformanceReport
The configuration data for an Ad Exchange performance report list.

PerformanceReportList
The configuration data for an Ad Exchange performance report list.

PerformanceReportListCall
Retrieves the authenticated user's list of performance metrics.

PerformanceReportMethods
A builder providing access to all methods supported on performanceReport resources. It is not used directly, but through the AdExchangeBuyer hub.

PretargetingConfig
There is no detailed description.

PretargetingConfigDeleteCall
Deletes an existing pretargeting config.

PretargetingConfigDimensions
Requests which allow one of these (width, height) pairs will match. All pairs must be supported ad dimensions.

PretargetingConfigExcludedPlacements
Requests containing any of these placements will not match.

PretargetingConfigGetCall
Gets a specific pretargeting configuration

PretargetingConfigInsertCall
Inserts a new pretargeting configuration.

PretargetingConfigList
There is no detailed description.

PretargetingConfigListCall
Retrieves a list of the authenticated user's pretargeting configurations.

PretargetingConfigMethods
A builder providing access to all methods supported on pretargetingConfig resources. It is not used directly, but through the AdExchangeBuyer hub.

PretargetingConfigPatchCall
Updates an existing pretargeting config. This method supports patch semantics.

PretargetingConfigPlacements
Requests containing any of these placements will match.

PretargetingConfigUpdateCall
Updates an existing pretargeting config.

PretargetingConfigVideoPlayerSizes
Video requests satisfying any of these player size constraints will match.

Price
There is no detailed description.

PricePerBuyer
Used to specify pricing rules for buyers. Each PricePerBuyer in a product can become [0,1] deals. To check if there is a PricePerBuyer for a particular buyer we look for the most specific matching rule - we first look for a rule matching the buyer and otherwise look for a matching rule where no buyer is set.

PrivateData
There is no detailed description.

Product
A product is segment of inventory that a seller wishes to sell. It is associated with certain terms and targeting information which helps buyer know more about the inventory. Each field in a product can have one of the following setting:

ProductGetCall
Gets the requested product by id.

ProductMethods
A builder providing access to all methods supported on product resources. It is not used directly, but through the AdExchangeBuyer hub.

ProductSearchCall
Gets the requested product.

Proposal
Represents a proposal in the marketplace. A proposal is the unit of negotiation between a seller and a buyer and contains deals which are served. Each field in a proposal can have one of the following setting:

ProposalGetCall
Get a proposal given its id

ProposalInsertCall
Create the given list of proposals

ProposalMethods
A builder providing access to all methods supported on proposal resources. It is not used directly, but through the AdExchangeBuyer hub.

ProposalPatchCall
Update the given proposal. This method supports patch semantics.

ProposalSearchCall
Search for proposals using pql query

ProposalSetupcompleteCall
Update the given proposal to indicate that setup has been completed.

ProposalUpdateCall
Update the given proposal

PublisherProfileApiProto
There is no detailed description.

PublisherProvidedForecast
This message carries publisher provided forecasting information.

PubprofileListCall
Gets the requested publisher profile(s) by publisher accountId.

PubprofileMethods
A builder providing access to all methods supported on pubprofile resources. It is not used directly, but through the AdExchangeBuyer hub.

Seller
There is no detailed description.

SharedTargeting
There is no detailed description.

TargetingValue
There is no detailed description.

TargetingValueCreativeSize
Next Id: 7

TargetingValueDayPartTargeting
There is no detailed description.

TargetingValueDayPartTargetingDayPart
There is no detailed description.

TargetingValueDemogAgeCriteria
There is no detailed description.

TargetingValueDemogGenderCriteria
There is no detailed description.

TargetingValueRequestPlatformTargeting
There is no detailed description.

TargetingValueSize
There is no detailed description.

UpdatePrivateAuctionProposalRequest
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.