Struct google_adexchangebuyer1d4::api::Product

pub struct Product {
    pub billed_buyer: Option<Buyer>,
    pub buyer: Option<Buyer>,
    pub creation_time_ms: Option<String>,
    pub creator_contacts: Option<Vec<ContactInformation>>,
    pub creator_role: Option<String>,
    pub delivery_control: Option<DeliveryControl>,
    pub flight_end_time_ms: Option<String>,
    pub flight_start_time_ms: Option<String>,
    pub has_creator_signed_off: Option<bool>,
    pub inventory_source: Option<String>,
    pub kind: Option<String>,
    pub labels: Option<Vec<MarketplaceLabel>>,
    pub last_update_time_ms: Option<String>,
    pub legacy_offer_id: Option<String>,
    pub marketplace_publisher_profile_id: Option<String>,
    pub name: Option<String>,
    pub private_auction_id: Option<String>,
    pub product_id: Option<String>,
    pub publisher_profile_id: Option<String>,
    pub publisher_provided_forecast: Option<PublisherProvidedForecast>,
    pub revision_number: Option<String>,
    pub seller: Option<Seller>,
    pub shared_targetings: Option<Vec<SharedTargeting>>,
    pub state: Option<String>,
    pub syndication_product: Option<String>,
    pub terms: Option<DealTerms>,
    pub web_property_code: Option<String>,
}

A product is segment of inventory that a seller wishes to sell. It is associated with certain terms and targeting information which helps buyer know more about the inventory. Each field in a product can have one of the following setting:

(readonly) - It is an error to try and set this field. (buyer-readonly) - Only the seller can set this field. (seller-readonly) - Only the buyer can set this field. (updatable) - The field is updatable at all times by either buyer or the seller.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get products (response)

search products (none)

Fields
billed_buyer: Option<Buyer>

The billed buyer corresponding to the buyer that created the offer. (readonly, except on create)

buyer: Option<Buyer>

The buyer that created the offer if this is a buyer initiated offer (readonly, except on create)

creation_time_ms: Option<String>

Creation time in ms. since epoch (readonly)

creator_contacts: Option<Vec<ContactInformation>>

Optional contact information for the creator of this product. (buyer-readonly)

creator_role: Option<String>

The role that created the offer. Set to BUYER for buyer initiated offers.

delivery_control: Option<DeliveryControl>

The set of fields around delivery control that are interesting for a buyer to see but are non-negotiable. These are set by the publisher. This message is assigned an id of 100 since some day we would want to model this as a protobuf extension.

flight_end_time_ms: Option<String>

The proposed end time for the deal (ms since epoch) (buyer-readonly)

flight_start_time_ms: Option<String>

Inventory availability dates. (times are in ms since epoch) The granularity is generally in the order of seconds. (buyer-readonly)

has_creator_signed_off: Option<bool>

If the creator has already signed off on the product, then the buyer can finalize the deal by accepting the product as is. When copying to a proposal, if any of the terms are changed, then auto_finalize is automatically set to false.

inventory_source: Option<String>

What exchange will provide this inventory (readonly, except on create).

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "adexchangebuyer#product".

labels: Option<Vec<MarketplaceLabel>>

Optional List of labels for the product (optional, buyer-readonly).

last_update_time_ms: Option<String>

Time of last update in ms. since epoch (readonly)

legacy_offer_id: Option<String>

Optional legacy offer id if this offer is a preferred deal offer.

marketplace_publisher_profile_id: Option<String>

Marketplace publisher profile Id. This Id differs from the regular publisher_profile_id in that 1. This is a new id, the old Id will be deprecated in 2017. 2. This id uniquely identifies a publisher profile by itself.

name: Option<String>

The name for this product as set by the seller. (buyer-readonly)

private_auction_id: Option<String>

Optional private auction id if this offer is a private auction offer.

product_id: Option<String>

The unique id for the product (readonly)

publisher_profile_id: Option<String>

Id of the publisher profile for a given seller. A (seller.account_id, publisher_profile_id) pair uniquely identifies a publisher profile. Buyers can call the PublisherProfiles::List endpoint to get a list of publisher profiles for a given seller.

publisher_provided_forecast: Option<PublisherProvidedForecast>

Publisher self-provided forecast information.

revision_number: Option<String>

The revision number of the product. (readonly)

seller: Option<Seller>

Information about the seller that created this product (readonly, except on create)

shared_targetings: Option<Vec<SharedTargeting>>

Targeting that is shared between the buyer and the seller. Each targeting criteria has a specified key and for each key there is a list of inclusion value or exclusion values. (buyer-readonly)

state: Option<String>

The state of the product. (buyer-readonly)

syndication_product: Option<String>

The syndication product associated with the deal. (readonly, except on create)

terms: Option<DealTerms>

The negotiable terms of the deal (buyer-readonly)

web_property_code: Option<String>

The web property code for the seller. This field is meant to be copied over as is when creating deals.