Struct google_adexchangebuyer1d4::Creative

pub struct Creative {
    pub product_categories: Option<Vec<i32>>,
    pub agency_id: Option<String>,
    pub ad_technology_providers: Option<CreativeAdTechnologyProviders>,
    pub attribute: Option<Vec<i32>>,
    pub restricted_categories: Option<Vec<i32>>,
    pub detected_domains: Option<Vec<String>>,
    pub height: Option<i32>,
    pub advertiser_id: Option<Vec<String>>,
    pub html_snippet: Option<String>,
    pub open_auction_status: Option<String>,
    pub buyer_creative_id: Option<String>,
    pub filtering_reasons: Option<CreativeFilteringReasons>,
    pub deals_status: Option<String>,
    pub ad_choices_destination_url: Option<String>,
    pub impression_tracking_url: Option<Vec<String>>,
    pub account_id: Option<i32>,
    pub advertiser_name: Option<String>,
    pub kind: Option<String>,
    pub video_url: Option<String>,
    pub video_vast_xml: Option<String>,
    pub click_through_url: Option<Vec<String>>,
    pub serving_restrictions: Option<Vec<CreativeServingRestrictions>>,
    pub corrections: Option<Vec<CreativeCorrections>>,
    pub languages: Option<Vec<String>>,
    pub width: Option<i32>,
    pub native_ad: Option<CreativeNativeAd>,
    pub version: Option<i32>,
    pub api_upload_timestamp: Option<String>,
    pub vendor_type: Option<Vec<i32>>,
    pub creative_status_identity_type: Option<String>,
    pub sensitive_categories: Option<Vec<i32>>,
}

A creative and its classification data.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

add deal creatives (none)

insert creatives (request|response)

list creatives (none)

list deals creatives (none)

remove deal creatives (none)

get creatives (response)

Fields
product_categories: Option<Vec<i32>>

Detected product categories, if any. Each category is represented by an integer as defined in ad-product-categories.txt. Read-only. This field should not be set in requests.

agency_id: Option<String>

The agency id for this creative.

ad_technology_providers: Option<CreativeAdTechnologyProviders>

no description provided

attribute: Option<Vec<i32>>

List of buyer selectable attributes for the ads that may be shown from this snippet. Each attribute is represented by an integer as defined in buyer-declarable-creative-attributes.txt.

restricted_categories: Option<Vec<i32>>

All restricted categories for the ads that may be shown from this snippet. Each category is represented by an integer as defined in the ad-restricted-categories.txt.

detected_domains: Option<Vec<String>>

Detected domains for this creative. Read-only. This field should not be set in requests.

height: Option<i32>

Ad height.

advertiser_id: Option<Vec<String>>

Detected advertiser id, if any. Read-only. This field should not be set in requests.

html_snippet: Option<String>

The HTML snippet that displays the ad when inserted in the web page. If set, videoURL, videoVastXML, and nativeAd should not be set.

open_auction_status: Option<String>

Top-level open auction status. Read-only. This field should not be set in requests. If disapproved, an entry for auctionType=OPEN_AUCTION (or ALL) in servingRestrictions will also exist. Note that this may be nuanced with other contextual restrictions, in which case it may be preferable to read from ServingRestrictions directly.

buyer_creative_id: Option<String>

A buyer-specific id identifying the creative in this ad.

filtering_reasons: Option<CreativeFilteringReasons>

The filtering reasons for the creative. Read-only. This field should not be set in requests.

deals_status: Option<String>

Top-level deals status. Read-only. This field should not be set in requests. If disapproved, an entry for auctionType=DIRECT_DEALS (or ALL) in servingRestrictions will also exist. Note that this may be nuanced with other contextual restrictions, in which case it may be preferable to read from servingRestrictions directly.

ad_choices_destination_url: Option<String>

The link to the Ad Preferences page. This is only supported for native ads.

impression_tracking_url: Option<Vec<String>>

The set of urls to be called to record an impression.

account_id: Option<i32>

Account id.

advertiser_name: Option<String>

The name of the company being advertised in the creative. The value provided must exist in the advertisers.txt file.

kind: Option<String>

Resource type.

video_url: Option<String>

The URL to fetch a video ad. If set, HTMLSnippet, videoVastXML, and nativeAd should not be set. Note, this is different from resource.native_ad.video_url above.

video_vast_xml: Option<String>

The contents of a VAST document for a video ad. This document should conform to the VAST 2.0 or 3.0 standard. If set, HTMLSnippet, videoURL, and nativeAd and should not be set.

click_through_url: Option<Vec<String>>

The set of destination urls for the snippet.

serving_restrictions: Option<Vec<CreativeServingRestrictions>>

The granular status of this ad in specific contexts. A context here relates to where something ultimately serves (for example, a physical location, a platform, an HTTPS vs HTTP request, or the type of auction). Read-only. This field should not be set in requests. See the examples in the Creatives guide for more details.

corrections: Option<Vec<CreativeCorrections>>

Shows any corrections that were applied to this creative. Read-only. This field should not be set in requests.

languages: Option<Vec<String>>

Detected languages for this creative. Read-only. This field should not be set in requests.

width: Option<i32>

Ad width.

native_ad: Option<CreativeNativeAd>

If nativeAd is set, HTMLSnippet, videoVastXML, and the videoURL outside of nativeAd should not be set. (The videoURL inside nativeAd can be set.)

version: Option<i32>

The version for this creative. Read-only. This field should not be set in requests.

api_upload_timestamp: Option<String>

The last upload timestamp of this creative if it was uploaded via API. Read-only. The value of this field is generated, and will be ignored for uploads. (formatted RFC 3339 timestamp).

vendor_type: Option<Vec<i32>>

List of vendor types for the ads that may be shown from this snippet. Each vendor type is represented by an integer as defined in vendors.txt.

creative_status_identity_type: Option<String>

Creative status identity type that the creative item applies to. Ad Exchange real-time bidding is migrating to the sizeless creative verification. Originally, Ad Exchange assigned creative verification status to a unique combination of a buyer creative ID and creative dimensions. Post-migration, a single verification status will be assigned at the buyer creative ID level. This field allows to distinguish whether a given creative status applies to a unique combination of a buyer creative ID and creative dimensions, or to a buyer creative ID as a whole.

sensitive_categories: Option<Vec<i32>>

Detected sensitive categories, if any. Each category is represented by an integer as defined in ad-sensitive-categories.txt. Read-only. This field should not be set in requests.