Struct google_adexchangebuyer1d4::CreativeNativeAd

pub struct CreativeNativeAd {
    pub body: Option<String>,
    pub advertiser: Option<String>,
    pub headline: Option<String>,
    pub image: Option<CreativeNativeAdImage>,
    pub star_rating: Option<f64>,
    pub click_tracking_url: Option<String>,
    pub call_to_action: Option<String>,
    pub app_icon: Option<CreativeNativeAdAppIcon>,
    pub logo: Option<CreativeNativeAdLogo>,
    pub impression_tracking_url: Option<Vec<String>>,
    pub price: Option<String>,
    pub store: Option<String>,
}

If nativeAd is set, HTMLSnippet and videoURL should not be set.

This type is not used in any activity, and only used as part of another schema.

Fields

body

A long description of the ad.

advertiser

no description provided

headline

A short title for the ad.

image

A large image.

star_rating

The app rating in the app store. Must be in the range [0-5].

click_tracking_url

The URL to use for click tracking.

call_to_action

A label for the button that the user is supposed to click.

app_icon

The app icon, for app download ads.

logo

A smaller image, for the advertiser logo.

impression_tracking_url

The URLs are called when the impression is rendered.

price

The price of the promoted app including the currency info.

store

The URL to the app store to purchase/download the promoted app.