Struct google_adexchangebuyer1d4::DealTerms

pub struct DealTerms {
    pub branding_type: Option<String>,
    pub cross_listed_external_deal_id_type: Option<String>,
    pub guaranteed_fixed_price_terms: Option<DealTermsGuaranteedFixedPriceTerms>,
    pub seller_time_zone: Option<String>,
    pub estimated_impressions_per_day: Option<String>,
    pub non_guaranteed_fixed_price_terms: Option<DealTermsNonGuaranteedFixedPriceTerms>,
    pub estimated_gross_spend: Option<Price>,
    pub rubicon_non_guaranteed_terms: Option<DealTermsRubiconNonGuaranteedTerms>,
    pub non_guaranteed_auction_terms: Option<DealTermsNonGuaranteedAuctionTerms>,
    pub description: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
branding_type: Option<String>

Visibilty of the URL in bid requests.

cross_listed_external_deal_id_type: Option<String>

Indicates that this ExternalDealId exists under at least two different AdxInventoryDeals. Currently, the only case that the same ExternalDealId will exist is programmatic cross sell case.

guaranteed_fixed_price_terms: Option<DealTermsGuaranteedFixedPriceTerms>

The terms for guaranteed fixed price deals.

seller_time_zone: Option<String>

For deals with Cost Per Day billing, defines the timezone used to mark the boundaries of a day (buyer-readonly)

estimated_impressions_per_day: Option<String>

Non-binding estimate of the impressions served per day Can be set by buyer or seller.

non_guaranteed_fixed_price_terms: Option<DealTermsNonGuaranteedFixedPriceTerms>

The terms for non-guaranteed fixed price deals.

estimated_gross_spend: Option<Price>

Non-binding estimate of the estimated gross spend for this deal Can be set by buyer or seller.

rubicon_non_guaranteed_terms: Option<DealTermsRubiconNonGuaranteedTerms>

The terms for rubicon non-guaranteed deals.

non_guaranteed_auction_terms: Option<DealTermsNonGuaranteedAuctionTerms>

The terms for non-guaranteed auction deals.

description: Option<String>

Description for the proposed terms of the deal.