Struct google_adexchangebuyer1d4::MarketplaceDeal

pub struct MarketplaceDeal {
    pub delivery_control: Option<DeliveryControl>,
    pub terms: Option<DealTerms>,
    pub flight_end_time_ms: Option<String>,
    pub shared_targetings: Option<Vec<SharedTargeting>>,
    pub deal_id: Option<String>,
    pub programmatic_creative_source: Option<String>,
    pub creative_safe_frame_compatibility: Option<String>,
    pub external_deal_id: Option<String>,
    pub last_update_time_ms: Option<String>,
    pub seller_contacts: Option<Vec<ContactInformation>>,
    pub kind: Option<String>,
    pub creative_pre_approval_policy: Option<String>,
    pub name: Option<String>,
    pub buyer_private_data: Option<PrivateData>,
    pub syndication_product: Option<String>,
    pub product_revision_number: Option<String>,
    pub web_property_code: Option<String>,
    pub proposal_id: Option<String>,
    pub flight_start_time_ms: Option<String>,
    pub deal_serving_metadata: Option<DealServingMetadata>,
    pub creation_time_ms: Option<String>,
    pub inventory_description: Option<String>,
    pub product_id: Option<String>,
}

A proposal can contain multiple deals. A deal contains the terms and targeting information that is used for serving.

This type is not used in any activity, and only used as part of another schema.

Fields
delivery_control: Option<DeliveryControl>

The set of fields around delivery control that are interesting for a buyer to see but are non-negotiable. These are set by the publisher. This message is assigned an id of 100 since some day we would want to model this as a protobuf extension.

terms: Option<DealTerms>

The negotiable terms of the deal. (updatable)

flight_end_time_ms: Option<String>

Proposed flight end time of the deal (ms since epoch) This will generally be stored in a granularity of a second. (updatable)

shared_targetings: Option<Vec<SharedTargeting>>

The shared targeting visible to buyers and sellers. Each shared targeting entity is AND'd together. (updatable)

deal_id: Option<String>

A unique deal-id for the deal (readonly).

programmatic_creative_source: Option<String>

Specifies the creative source for programmatic deals, PUBLISHER means creative is provided by seller and ADVERTISR means creative is provided by buyer. (buyer-readonly)

creative_safe_frame_compatibility: Option<String>

Specifies whether the creative is safeFrame compatible (buyer-readonly)

external_deal_id: Option<String>

The external deal id assigned to this deal once the deal is finalized. This is the deal-id that shows up in serving/reporting etc. (readonly)

last_update_time_ms: Option<String>

The time (ms since epoch) when the deal was last updated. (readonly)

seller_contacts: Option<Vec<ContactInformation>>

Optional Seller contact information for the deal (buyer-readonly)

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "adexchangebuyer#marketplaceDeal".

creative_pre_approval_policy: Option<String>

Specifies the creative pre-approval policy (buyer-readonly)

name: Option<String>

The name of the deal. (updatable)

buyer_private_data: Option<PrivateData>

Buyer private data (hidden from seller).

syndication_product: Option<String>

The syndication product associated with the deal. (readonly, except on create)

product_revision_number: Option<String>

The revision number of the product that the deal was created from (readonly, except on create)

web_property_code: Option<String>

no description provided

proposal_id: Option<String>

no description provided

flight_start_time_ms: Option<String>

Proposed flight start time of the deal (ms since epoch) This will generally be stored in a granularity of a second. (updatable)

deal_serving_metadata: Option<DealServingMetadata>

Metadata about the serving status of this deal (readonly, writes via custom actions)

creation_time_ms: Option<String>

The time (ms since epoch) of the deal creation. (readonly)

inventory_description: Option<String>

Description for the deal terms. (updatable)

product_id: Option<String>

The product-id from which this deal was created. (readonly, except on create)