Struct google_adexchangebuyer1d4::MarketplaceNote

pub struct MarketplaceNote {
    pub kind: Option<String>,
    pub proposal_revision_number: Option<String>,
    pub deal_id: Option<String>,
    pub note: Option<String>,
    pub proposal_id: Option<String>,
    pub creator_role: Option<String>,
    pub note_id: Option<String>,
    pub timestamp_ms: Option<String>,
}

A proposal is associated with a bunch of notes which may optionally be associated with a deal and/or revision number.

This type is not used in any activity, and only used as part of another schema.

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "adexchangebuyer#marketplaceNote".

proposal_revision_number

If the note is associated with a proposal revision number, then store that here. (readonly, except on create)

deal_id

Notes can optionally be associated with a deal. (readonly, except on create)

note

The actual note to attach. (readonly, except on create)

proposal_id

The proposalId that a note is attached to. (readonly)

creator_role

The role of the person (buyer/seller) creating the note. (readonly)

note_id

The unique id for the note. (readonly)

timestamp_ms

The timestamp (ms since epoch) that this note was created. (readonly)