Struct google_adexchangebuyer1d4::PretargetingConfigMethods

pub struct PretargetingConfigMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on pretargetingConfig resources. It is not used directly, but through the AdExchangeBuyer hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adexchangebuyer1d4 as adexchangebuyer1d4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adexchangebuyer1d4::AdExchangeBuyer;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdExchangeBuyer::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.pretargeting_config();

Methods

impl<'a, C, A> PretargetingConfigMethods<'a, C, A>

fn insert(&self, request: PretargetingConfig, account_id: &str) -> PretargetingConfigInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts a new pretargeting configuration.

Arguments

request - No description provided.

accountId - The account id to insert the pretargeting config for.

fn list(&self, account_id: &str) -> PretargetingConfigListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of the authenticated user's pretargeting configurations.

Arguments

accountId - The account id to get the pretargeting configs for.

fn update(&self, request: PretargetingConfig, account_id: &str, config_id: &str) -> PretargetingConfigUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing pretargeting config.

Arguments

request - No description provided.

accountId - The account id to update the pretargeting config for.

configId - The specific id of the configuration to update.

fn patch(&self, request: PretargetingConfig, account_id: &str, config_id: &str) -> PretargetingConfigPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing pretargeting config. This method supports patch semantics.

Arguments

request - No description provided.

accountId - The account id to update the pretargeting config for.

configId - The specific id of the configuration to update.

fn delete(&self, account_id: &str, config_id: &str) -> PretargetingConfigDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an existing pretargeting config.

Arguments

accountId - The account id to delete the pretargeting config for.

configId - The specific id of the configuration to delete.

fn get(&self, account_id: &str, config_id: &str) -> PretargetingConfigGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a specific pretargeting configuration

Arguments

accountId - The account id to get the pretargeting config for.

configId - The specific id of the configuration to retrieve.