Struct google_adexchangebuyer1d4::Product

pub struct Product {
    pub creator_contacts: Option<Vec<ContactInformation>>,
    pub delivery_control: Option<DeliveryControl>,
    pub shared_targetings: Option<Vec<SharedTargeting>>,
    pub terms: Option<DealTerms>,
    pub private_auction_id: Option<String>,
    pub flight_end_time_ms: Option<String>,
    pub labels: Option<Vec<MarketplaceLabel>>,
    pub inventory_source: Option<String>,
    pub publisher_profile_id: Option<String>,
    pub last_update_time_ms: Option<String>,
    pub revision_number: Option<String>,
    pub legacy_offer_id: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub syndication_product: Option<String>,
    pub seller: Option<Seller>,
    pub web_property_code: Option<String>,
    pub state: Option<String>,
    pub has_creator_signed_off: Option<bool>,
    pub flight_start_time_ms: Option<String>,
    pub creation_time_ms: Option<String>,
    pub product_id: Option<String>,
}

A product is segment of inventory that a seller wishes to sell. It is associated with certain terms and targeting information which helps buyer know more about the inventory. Each field in a product can have one of the following setting:

(readonly) - It is an error to try and set this field. (buyer-readonly) - Only the seller can set this field. (seller-readonly) - Only the buyer can set this field. (updatable) - The field is updatable at all times by either buyer or the seller.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

search products (none)

get products (response)

Fields

creator_contacts

Optional contact information for the creator of this product. (buyer-readonly)

delivery_control

The set of fields around delivery control that are interesting for a buyer to see but are non-negotiable. These are set by the publisher. This message is assigned an id of 100 since some day we would want to model this as a protobuf extension.

shared_targetings

Targeting that is shared between the buyer and the seller. Each targeting criteria has a specified key and for each key there is a list of inclusion value or exclusion values. (buyer-readonly)

terms

The negotiable terms of the deal (buyer-readonly)

private_auction_id

Optional private auction id if this offer is a private auction offer.

flight_end_time_ms

The proposed end time for the deal (ms since epoch) (buyer-readonly)

labels

Optional List of labels for the product (optional, buyer-readonly).

inventory_source

What exchange will provide this inventory (readonly, except on create).

publisher_profile_id

Id of the publisher profile for a given seller. A (seller.account_id, publisher_profile_id) pair uniquely identifies a publisher profile. Buyers can call the PublisherProfiles::List endpoint to get a list of publisher profiles for a given seller.

last_update_time_ms

Time of last update in ms. since epoch (readonly)

revision_number

The revision number of the product. (readonly)

legacy_offer_id

Optional legacy offer id if this offer is a preferred deal offer.

kind

Identifies what kind of resource this is. Value: the fixed string "adexchangebuyer#product".

name

The name for this product as set by the seller. (buyer-readonly)

syndication_product

The syndication product associated with the deal. (readonly, except on create)

seller

Information about the seller that created this product (readonly, except on create)

web_property_code

The web property code for the seller. This field is meant to be copied over as is when creating deals.

state

The state of the product. (buyer-readonly)

has_creator_signed_off

If the creator has already signed off on the product, then the buyer can finalize the deal by accepting the product as is. When copying to a proposal, if any of the terms are changed, then auto_finalize is automatically set to false.

flight_start_time_ms

Inventory availability dates. (times are in ms since epoch) The granularity is generally in the order of seconds. (buyer-readonly)

creation_time_ms

Creation time in ms. since epoch (readonly)

product_id

The unique id for the product (readonly)