Struct google_adexchangebuyer1d4::Proposal

pub struct Proposal {
    pub private_auction_id: Option<String>,
    pub labels: Option<Vec<MarketplaceLabel>>,
    pub originator_role: Option<String>,
    pub seller_contacts: Option<Vec<ContactInformation>>,
    pub negotiation_id: Option<String>,
    pub inventory_source: Option<String>,
    pub has_seller_signed_off: Option<bool>,
    pub buyer: Option<Buyer>,
    pub has_buyer_signed_off: Option<bool>,
    pub revision_number: Option<String>,
    pub kind: Option<String>,
    pub proposal_state: Option<String>,
    pub name: Option<String>,
    pub revision_time_ms: Option<String>,
    pub buyer_private_data: Option<PrivateData>,
    pub billed_buyer: Option<Buyer>,
    pub seller: Option<Seller>,
    pub proposal_id: Option<String>,
    pub is_renegotiating: Option<bool>,
    pub is_setup_complete: Option<bool>,
    pub last_updater_or_commentor_role: Option<String>,
    pub buyer_contacts: Option<Vec<ContactInformation>>,
}

Represents a proposal in the marketplace. A proposal is the unit of negotiation between a seller and a buyer and contains deals which are served. Each field in a proposal can have one of the following setting:

(readonly) - It is an error to try and set this field. (buyer-readonly) - Only the seller can set this field. (seller-readonly) - Only the buyer can set this field. (updatable) - The field is updatable at all times by either buyer or the seller.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update proposals (request|response)

search proposals (none)

insert proposals (none)

setupcomplete proposals (none)

patch proposals (request|response)

get proposals (response)

Fields
private_auction_id: Option<String>

Optional private auction id if this proposal is a private auction proposal.

labels: Option<Vec<MarketplaceLabel>>

List of labels associated with the proposal. (readonly)

originator_role: Option<String>

Indicates whether the buyer/seller created the proposal.(readonly)

seller_contacts: Option<Vec<ContactInformation>>

Optional contact information of the seller (buyer-readonly).

negotiation_id: Option<String>

Optional negotiation id if this proposal is a preferred deal proposal.

inventory_source: Option<String>

What exchange will provide this inventory (readonly, except on create).

has_seller_signed_off: Option<bool>

When an proposal is in an accepted state, indicates whether the buyer has signed off Once both sides have signed off on a deal, the proposal can be finalized by the seller. (buyer-readonly)

buyer: Option<Buyer>

Reference to the buyer on the proposal. (readonly, except on create)

has_buyer_signed_off: Option<bool>

When an proposal is in an accepted state, indicates whether the buyer has signed off. Once both sides have signed off on a deal, the proposal can be finalized by the seller. (seller-readonly)

revision_number: Option<String>

The revision number for the proposal (readonly).

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "adexchangebuyer#proposal".

proposal_state: Option<String>

The current state of the proposal. (readonly)

name: Option<String>

The name for the proposal (updatable)

revision_time_ms: Option<String>

The time (ms since epoch) when the proposal was last revised (readonly).

buyer_private_data: Option<PrivateData>

Private data for buyer. (hidden from seller).

billed_buyer: Option<Buyer>

Reference to the buyer that will get billed for this proposal. (readonly)

seller: Option<Seller>

Reference to the seller on the proposal. (readonly, except on create)

proposal_id: Option<String>

The unique id of the proposal. (readonly).

is_renegotiating: Option<bool>

True if the proposal is being renegotiated (readonly).

is_setup_complete: Option<bool>

True, if the buyside inventory setup is complete for this proposal. (readonly, except via OrderSetupCompleted action)

last_updater_or_commentor_role: Option<String>

The role of the last user that either updated the proposal or left a comment. (readonly)

buyer_contacts: Option<Vec<ContactInformation>>

Optional contact information of the buyer. (seller-readonly)