Struct google_adexchangebuyer1d4::ProposalMethods

pub struct ProposalMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on proposal resources. It is not used directly, but through the AdExchangeBuyer hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adexchangebuyer1d4 as adexchangebuyer1d4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adexchangebuyer1d4::AdExchangeBuyer;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdExchangeBuyer::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `insert(...)`, `patch(...)`, `search(...)` and `update(...)`
// to build up your call.
let rb = hub.proposals();

Methods

impl<'a, C, A> ProposalMethods<'a, C, A>

fn search(&self) -> ProposalSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Search for proposals using pql query

fn insert(&self, request: CreateOrdersRequest) -> ProposalInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create the given list of proposals

Arguments

request - No description provided.

fn update(&self, request: Proposal, proposal_id: &str, revision_number: &str, update_action: &str) -> ProposalUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update the given proposal

Arguments

request - No description provided.

proposalId - The proposal id to update.

revisionNumber - The last known revision number to update. If the head revision in the marketplace database has since changed, an error will be thrown. The caller should then fetch the latest proposal at head revision and retry the update at that revision.

updateAction - The proposed action to take on the proposal.

fn patch(&self, request: Proposal, proposal_id: &str, revision_number: &str, update_action: &str) -> ProposalPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update the given proposal. This method supports patch semantics.

Arguments

request - No description provided.

proposalId - The proposal id to update.

revisionNumber - The last known revision number to update. If the head revision in the marketplace database has since changed, an error will be thrown. The caller should then fetch the latest proposal at head revision and retry the update at that revision.

updateAction - The proposed action to take on the proposal.

fn get(&self, proposal_id: &str) -> ProposalGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get a proposal given its id

Arguments

proposalId - Id of the proposal to retrieve.