Struct google_adexchangebuyer1d4::PublisherProfileApiProto

pub struct PublisherProfileApiProto {
    pub google_plus_link: Option<String>,
    pub media_kit_link: Option<String>,
    pub name: Option<String>,
    pub publisher_provided_forecast: Option<PublisherProvidedForecast>,
    pub overview: Option<String>,
    pub buyer_pitch_statement: Option<String>,
    pub is_parent: Option<bool>,
    pub top_headlines: Option<Vec<String>>,
    pub audience: Option<String>,
    pub publisher_profile_id: Option<String>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub programmatic_contact: Option<ContactInformation>,
    pub publisher_domains: Option<Vec<String>>,
    pub exchange: Option<String>,
    pub rate_card_info_link: Option<String>,
    pub is_published: Option<bool>,
    pub seller: Option<Seller>,
    pub direct_contact: Option<ContactInformation>,
    pub sample_page_link: Option<String>,
    pub state: Option<String>,
    pub logo_url: Option<String>,
    pub profile_id: Option<i32>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

google_plus_link

Link to publisher's Google+ page.

media_kit_link

The url for additional marketing and sales materials.

name

no description provided

publisher_provided_forecast

Publisher provided forecasting information.

overview

Publisher provided overview.

buyer_pitch_statement

A pitch statement for the buyer

is_parent

True, if this is the parent profile, which represents all domains owned by the publisher.

top_headlines

Publisher provided key metrics and rankings.

audience

Publisher provided info on its audience.

publisher_profile_id

Unique Id for publisher profile.

account_id

The account id of the seller.

kind

Identifies what kind of resource this is. Value: the fixed string "adexchangebuyer#publisherProfileApiProto".

programmatic_contact

Programmatic contact for the publisher profile.

publisher_domains

The list of domains represented in this publisher profile. Empty if this is a parent profile.

exchange

Exchange where this publisher profile is from. E.g. AdX, Rubicon etc...

rate_card_info_link

Link to publisher rate card

is_published

True, if this profile is published. Deprecated for state.

seller

Seller of the publisher profile.

direct_contact

Direct contact for the publisher profile.

sample_page_link

Link for a sample content page.

state

State of the publisher profile.

logo_url

The url to the logo for the publisher.

profile_id

The pair of (seller.account_id, profile_id) uniquely identifies a publisher profile for a given publisher.