Struct google_adexchangebuyer1d4::PubprofileMethods

pub struct PubprofileMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on pubprofile resources. It is not used directly, but through the AdExchangeBuyer hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adexchangebuyer1d4 as adexchangebuyer1d4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adexchangebuyer1d4::AdExchangeBuyer;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdExchangeBuyer::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `list(...)`
// to build up your call.
let rb = hub.pubprofiles();

Methods

impl<'a, C, A> PubprofileMethods<'a, C, A>

fn list(&self, account_id: i32) -> PubprofileListCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the requested publisher profile(s) by publisher accountId.

Arguments

accountId - The accountId of the publisher to get profiles for.