Crate google_adexchangebuyer1d4

This documentation was generated from Ad Exchange Buyer crate version 0.1.15+20160831, where 20160831 is the exact revision of the adexchangebuyer:v1.4 schema built by the mako code generator v0.1.15.

Everything else about the Ad Exchange Buyer v1d4 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

accounts

get, list, patch and update

billing info

get and list

budget

get, patch and update

creatives

add deal, get, insert, list and remove deal

marketplacedeals

delete, insert, list and update

marketplacenotes

insert and list

marketplaceprivateauction

updateproposal

performance report

list

pretargeting config

delete, get, insert, list, patch and update

products

get and search

proposals

get, insert, patch, search, setupcomplete and update

pubprofiles

list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.proposals().update(...).doit()
let r = hub.proposals().search(...).doit()
let r = hub.proposals().insert(...).doit()
let r = hub.proposals().setupcomplete(...).doit()
let r = hub.proposals().patch(...).doit()
let r = hub.proposals().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-adexchangebuyer1d4 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adexchangebuyer1d4 as adexchangebuyer1d4;
use adexchangebuyer1d4::Proposal;
use adexchangebuyer1d4::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adexchangebuyer1d4::AdExchangeBuyer;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdExchangeBuyer::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Proposal::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.proposals().update(req, "proposalId", "revisionNumber", "updateAction")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Account
Configuration data for an Ad Exchange buyer account.

AccountBidderLocation
Your bidder locations that have distinct URLs.

AccountGetCall
Gets one account by ID.

AccountListCall
Retrieves the authenticated user's list of accounts.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the AdExchangeBuyer hub.

AccountPatchCall
Updates an existing account. This method supports patch semantics.

AccountUpdateCall
Updates an existing account.

AccountsList
An account feed lists Ad Exchange buyer accounts that the user has access to. Each entry in the feed corresponds to a single buyer account.

AdExchangeBuyer
Central instance to access all AdExchangeBuyer related resource activities

AddOrderDealsRequest
There is no detailed description.

AddOrderDealsResponse
There is no detailed description.

AddOrderNotesRequest
There is no detailed description.

AddOrderNotesResponse
There is no detailed description.

BillingInfo
The configuration data for an Ad Exchange billing info.

BillingInfoGetCall
Returns the billing information for one account specified by account ID.

BillingInfoList
A billing info feed lists Billing Info the Ad Exchange buyer account has access to. Each entry in the feed corresponds to a single billing info.

BillingInfoListCall
Retrieves a list of billing information for all accounts of the authenticated user.

BillingInfoMethods
A builder providing access to all methods supported on billingInfo resources. It is not used directly, but through the AdExchangeBuyer hub.

Budget
The configuration data for Ad Exchange RTB - Budget API.

BudgetGetCall
Returns the budget information for the adgroup specified by the accountId and billingId.

BudgetMethods
A builder providing access to all methods supported on budget resources. It is not used directly, but through the AdExchangeBuyer hub.

BudgetPatchCall
Updates the budget amount for the budget of the adgroup specified by the accountId and billingId, with the budget amount in the request. This method supports patch semantics.

BudgetUpdateCall
Updates the budget amount for the budget of the adgroup specified by the accountId and billingId, with the budget amount in the request.

Buyer
There is no detailed description.

ContactInformation
There is no detailed description.

CreateOrdersRequest
There is no detailed description.

CreateOrdersResponse
There is no detailed description.

Creative
A creative and its classification data.

CreativeAddDealCall
Add a deal id association for the creative.

CreativeCorrections
Shows any corrections that were applied to this creative. Read-only. This field should not be set in requests.

CreativeCorrectionsContexts
All known serving contexts containing serving status information.

CreativeFilteringReasons
The filtering reasons for the creative. Read-only. This field should not be set in requests.

CreativeFilteringReasonsReasons
The filtering reasons.

CreativeGetCall
Gets the status for a single creative. A creative will be available 30-40 minutes after submission.

CreativeInsertCall
Submit a new creative.

CreativeListCall
Retrieves a list of the authenticated user's active creatives. A creative will be available 30-40 minutes after submission.

CreativeMethods
A builder providing access to all methods supported on creative resources. It is not used directly, but through the AdExchangeBuyer hub.

CreativeNativeAd
If nativeAd is set, HTMLSnippet and videoURL should not be set.

CreativeNativeAdAppIcon
The app icon, for app download ads.

CreativeNativeAdImage
A large image.

CreativeNativeAdLogo
A smaller image, for the advertiser logo.

CreativeRemoveDealCall
Remove a deal id associated with the creative.

CreativeServingRestrictions
The granular status of this ad in specific contexts. A context here relates to where something ultimately serves (for example, a physical location, a platform, an HTTPS vs HTTP request, or the type of auction). Read-only. This field should not be set in requests.

CreativeServingRestrictionsContexts
All known contexts/restrictions.

CreativeServingRestrictionsDisapprovalReasons
The reasons for disapproval within this restriction, if any. Note that not all disapproval reasons may be categorized, so it is possible for the creative to have a status of DISAPPROVED or CONDITIONALLY_APPROVED with an empty list for disapproval_reasons. In this case, please reach out to your TAM to help debug the issue.

CreativesList
The creatives feed lists the active creatives for the Ad Exchange buyer accounts that the user has access to. Each entry in the feed corresponds to a single creative.

DealServingMetadata
There is no detailed description.

DealServingMetadataDealPauseStatus
Tracks which parties (if any) have paused a deal. The deal is considered paused if has_buyer_paused || has_seller_paused. Each of the has_buyer_paused or the has_seller_paused bits can be set independently.

DealTerms
There is no detailed description.

DealTermsGuaranteedFixedPriceTerms
There is no detailed description.

DealTermsGuaranteedFixedPriceTermsBillingInfo
There is no detailed description.

DealTermsNonGuaranteedAuctionTerms
There is no detailed description.

DealTermsNonGuaranteedFixedPriceTerms
There is no detailed description.

DealTermsRubiconNonGuaranteedTerms
There is no detailed description.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DeleteOrderDealsRequest
There is no detailed description.

DeleteOrderDealsResponse
There is no detailed description.

DeliveryControl
There is no detailed description.

DeliveryControlFrequencyCap
There is no detailed description.

Dimension
This message carries publisher provided breakdown. E.g. {dimension_type: 'COUNTRY', [{dimension_value: {id: 1, name: 'US'}}, {dimension_value: {id: 2, name: 'UK'}}]}

DimensionDimensionValue
Value of the dimension.

EditAllOrderDealsRequest
There is no detailed description.

EditAllOrderDealsResponse
There is no detailed description.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

GetOffersResponse
There is no detailed description.

GetOrderDealsResponse
There is no detailed description.

GetOrderNotesResponse
There is no detailed description.

GetOrdersResponse
There is no detailed description.

GetPublisherProfilesByAccountIdResponse
There is no detailed description.

MarketplaceDeal
A proposal can contain multiple deals. A deal contains the terms and targeting information that is used for serving.

MarketplaceDealParty
There is no detailed description.

MarketplaceLabel
There is no detailed description.

MarketplaceNote
A proposal is associated with a bunch of notes which may optionally be associated with a deal and/or revision number.

MarketplacedealDeleteCall
Delete the specified deals from the proposal

MarketplacedealInsertCall
Add new deals for the specified proposal

MarketplacedealListCall
List all the deals for a given proposal

MarketplacedealMethods
A builder providing access to all methods supported on marketplacedeal resources. It is not used directly, but through the AdExchangeBuyer hub.

MarketplacedealUpdateCall
Replaces all the deals in the proposal with the passed in deals

MarketplacenoteInsertCall
Add notes to the proposal

MarketplacenoteListCall
Get all the notes associated with a proposal

MarketplacenoteMethods
A builder providing access to all methods supported on marketplacenote resources. It is not used directly, but through the AdExchangeBuyer hub.

MarketplaceprivateauctionMethods
A builder providing access to all methods supported on marketplaceprivateauction resources. It is not used directly, but through the AdExchangeBuyer hub.

MarketplaceprivateauctionUpdateproposalCall
Update a given private auction proposal

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

PerformanceReport
The configuration data for an Ad Exchange performance report list.

PerformanceReportList
The configuration data for an Ad Exchange performance report list.

PerformanceReportListCall
Retrieves the authenticated user's list of performance metrics.

PerformanceReportMethods
A builder providing access to all methods supported on performanceReport resources. It is not used directly, but through the AdExchangeBuyer hub.

PretargetingConfig
There is no detailed description.

PretargetingConfigDeleteCall
Deletes an existing pretargeting config.

PretargetingConfigDimensions
Requests which allow one of these (width, height) pairs will match. All pairs must be supported ad dimensions.

PretargetingConfigExcludedPlacements
Requests containing any of these placements will not match.

PretargetingConfigGetCall
Gets a specific pretargeting configuration

PretargetingConfigInsertCall
Inserts a new pretargeting configuration.

PretargetingConfigList
There is no detailed description.

PretargetingConfigListCall
Retrieves a list of the authenticated user's pretargeting configurations.

PretargetingConfigMethods
A builder providing access to all methods supported on pretargetingConfig resources. It is not used directly, but through the AdExchangeBuyer hub.

PretargetingConfigPatchCall
Updates an existing pretargeting config. This method supports patch semantics.

PretargetingConfigPlacements
Requests containing any of these placements will match.

PretargetingConfigUpdateCall
Updates an existing pretargeting config.

PretargetingConfigVideoPlayerSizes
Video requests satisfying any of these player size constraints will match.

Price
There is no detailed description.

PricePerBuyer
Used to specify pricing rules for buyers/advertisers. Each PricePerBuyer in an product can become [0,1] deals. To check if there is a PricePerBuyer for a particular buyer or buyer/advertiser pair, we look for the most specific matching rule - we first look for a rule matching the buyer and advertiser, next a rule with the buyer but an empty advertiser list, and otherwise look for a matching rule where no buyer is set.

PrivateData
There is no detailed description.

Product
A product is segment of inventory that a seller wishes to sell. It is associated with certain terms and targeting information which helps buyer know more about the inventory. Each field in a product can have one of the following setting:

ProductGetCall
Gets the requested product by id.

ProductMethods
A builder providing access to all methods supported on product resources. It is not used directly, but through the AdExchangeBuyer hub.

ProductSearchCall
Gets the requested product.

Proposal
Represents a proposal in the marketplace. A proposal is the unit of negotiation between a seller and a buyer and contains deals which are served. Each field in a proposal can have one of the following setting:

ProposalGetCall
Get a proposal given its id

ProposalInsertCall
Create the given list of proposals

ProposalMethods
A builder providing access to all methods supported on proposal resources. It is not used directly, but through the AdExchangeBuyer hub.

ProposalPatchCall
Update the given proposal. This method supports patch semantics.

ProposalSearchCall
Search for proposals using pql query

ProposalSetupcompleteCall
Update the given proposal to indicate that setup has been completed.

ProposalUpdateCall
Update the given proposal

PublisherProfileApiProto
There is no detailed description.

PublisherProvidedForecast
This message carries publisher provided forecasting information.

PubprofileListCall
Gets the requested publisher profile(s) by publisher accountId.

PubprofileMethods
A builder providing access to all methods supported on pubprofile resources. It is not used directly, but through the AdExchangeBuyer hub.

Seller
There is no detailed description.

SharedTargeting
There is no detailed description.

TargetingValue
There is no detailed description.

TargetingValueCreativeSize
There is no detailed description.

TargetingValueDayPartTargeting
There is no detailed description.

TargetingValueDayPartTargetingDayPart
There is no detailed description.

TargetingValueSize
There is no detailed description.

UpdatePrivateAuctionProposalRequest
There is no detailed description.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.