Struct google_adexchangebuyer1d4::Proposal

pub struct Proposal {
    pub labels: Option<Vec<MarketplaceLabel>>,
    pub originator_role: Option<String>,
    pub seller_contacts: Option<Vec<ContactInformation>>,
    pub buyer: Option<Buyer>,
    pub inventory_source: Option<String>,
    pub has_seller_signed_off: Option<bool>,
    pub last_updater_role: Option<String>,
    pub has_buyer_signed_off: Option<bool>,
    pub revision_number: Option<String>,
    pub kind: Option<String>,
    pub proposal_state: Option<String>,
    pub name: Option<String>,
    pub revision_time_ms: Option<String>,
    pub buyer_private_data: Option<PrivateData>,
    pub billed_buyer: Option<Buyer>,
    pub seller: Option<Seller>,
    pub proposal_id: Option<String>,
    pub is_renegotiating: Option<bool>,
    pub is_setup_complete: Option<bool>,
    pub last_updater_or_commentor_role: Option<String>,
    pub buyer_contacts: Option<Vec<ContactInformation>>,
}

Represents a proposal in the marketplace. A proposal is the unit of negotiation between a seller and a buyer and contains deals which are served. Each field in a proposal can have one of the following setting:

(readonly) - It is an error to try and set this field. (buyer-readonly) - Only the seller can set this field. (seller-readonly) - Only the buyer can set this field. (updatable) - The field is updatable at all times by either buyer or the seller.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

search proposals (none)

patch proposals (request|response)

insert proposals (none)

get proposals (response)

update proposals (request|response)

Fields

labels

List of labels associated with the proposal. (readonly)

originator_role

Indicates whether the buyer/seller created the proposal.(readonly)

seller_contacts

Optional contact information for the seller (buyer-readonly).

buyer

Reference to the buyer on the proposal. (readonly, except on create)

inventory_source

What exchange will provide this inventory (readonly, except on create).

has_seller_signed_off

When an proposal is in an accepted state, indicates whether the buyer has signed off Once both sides have signed off on a deal, the proposal can be finalized by the seller. (buyer-readonly)

last_updater_role

no description provided

has_buyer_signed_off

When an proposal is in an accepted state, indicates whether the buyer has signed off Once both sides have signed off on a deal, the proposal can be finalized by the seller. (seller-readonly)

revision_number

The revision number for the proposal (readonly).

kind

Identifies what kind of resource this is. Value: the fixed string "adexchangebuyer#proposal".

proposal_state

The current state of the proposal. (readonly)

name

The name for the proposal (updatable)

revision_time_ms

The time (ms since epoch) when the proposal was last revised (readonly).

buyer_private_data

Private data for buyer. (hidden from seller).

billed_buyer

Reference to the buyer that will get billed for this proposal. (readonly)

seller

Reference to the seller on the proposal. (readonly, except on create)

proposal_id

The unique id of the proposal. (readonly).

is_renegotiating

True if the proposal is being renegotiated (readonly).

is_setup_complete

True, if the buyside inventory setup is complete for this proposal. (readonly)

last_updater_or_commentor_role

The role of the last user that either updated the proposal or left a comment. (readonly)

buyer_contacts

Optional contact information fort the buyer. (seller-readonly)