Struct google_adexchangebuyer1d4::PublisherProfileApiProto

pub struct PublisherProfileApiProto {
    pub google_plus_link: Option<String>,
    pub media_kit_link: Option<String>,
    pub name: Option<String>,
    pub publisher_provided_forecast: Option<PublisherProvidedForecast>,
    pub overview: Option<String>,
    pub buyer_pitch_statement: Option<String>,
    pub is_parent: Option<bool>,
    pub top_headlines: Option<Vec<String>>,
    pub audience: Option<String>,
    pub publisher_profile_id: Option<String>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub programmatic_contact: Option<String>,
    pub publisher_domains: Option<Vec<String>>,
    pub exchange: Option<String>,
    pub rate_card_info_link: Option<String>,
    pub is_published: Option<bool>,
    pub seller: Option<Seller>,
    pub direct_contact: Option<String>,
    pub sample_page_link: Option<String>,
    pub state: Option<String>,
    pub logo_url: Option<String>,
    pub profile_id: Option<i32>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
google_plus_link: Option<String>

Link to publisher's Google+ page.

media_kit_link: Option<String>

The url for additional marketing and sales materials.

name: Option<String>

no description provided

publisher_provided_forecast: Option<PublisherProvidedForecast>

Publisher provided forecasting information.

overview: Option<String>

Publisher provided overview.

buyer_pitch_statement: Option<String>

A pitch statement for the buyer

is_parent: Option<bool>

True, if this is the parent profile, which represents all domains owned by the publisher.

top_headlines: Option<Vec<String>>

Publisher provided key metrics and rankings.

audience: Option<String>

Publisher provided info on its audience.

publisher_profile_id: Option<String>

Unique Id for publisher profile.

account_id: Option<String>

The account id of the seller.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "adexchangebuyer#publisherProfileApiProto".

programmatic_contact: Option<String>

Programmatic contact for the publisher profile.

publisher_domains: Option<Vec<String>>

The list of domains represented in this publisher profile. Empty if this is a parent profile.

exchange: Option<String>

Exchange where this publisher profile is from. E.g. AdX, Rubicon etc...

rate_card_info_link: Option<String>

Link to publisher rate card

is_published: Option<bool>

True, if this profile is published. Deprecated for state.

seller: Option<Seller>

Seller of the publisher profile.

direct_contact: Option<String>

Direct contact for the publisher profile.

sample_page_link: Option<String>

Link for a sample content page.

state: Option<String>

State of the publisher profile.

logo_url: Option<String>

The url to the logo for the publisher.

profile_id: Option<i32>

The pair of (seller.account_id, profile_id) uniquely identifies a publisher profile for a given publisher.